Move getDebugInfo from tgs to debug page
- getDebugInfo can never return if the target tab is in a closed tab
  group.  So, add a timeout to the Promize.all to allow the page to
  render if some tabs are unresponsive
- Keep debug function local to the debug page
